Added logo for Novocoin NVC token in info.json file. The logo has been successfully added for inclusion in the PancakeSwap token list.
I've added the Novocoin NVC token to the list of PancakeSwap tokens by submitting the info.json file and the link to the token's logo. This file contains the information needed to integrate the token on PancakeSwap, including the name, symbol, contract address and link to the logo image. The token is of the BEP20 type and uses the Binance Smart Chain. The logo is hosted at https://ibb.co/stFv8dW.